+++ Source/SwiftLintFramework/Rules/Style/IdentifierNameRule.swift
@@ -63,8 +63,7 @@ public struct IdentifierNameRule: ASTRule, ConfigurationProviderRule {
                 }
             }
 
-            let requiresCaseCheck = configuration.validatesStartWithLowercase || isFunction
-            if requiresCaseCheck &&
+            if configuration.validatesStartWithLowercase &&
                 kind != .varStatic && name.isViolatingCase && !name.isOperator {
                 let reason = "\(type) name should start with a lowercase character: '\(name)'"
                 return [