+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Master
+## 0.40.2: Demo Unit
 
 #### Breaking